Reorganize file structure to be more clear
Move files which are common to both the learner and admin side of things
to the `common/` directory. Also move some ContentBlock types into their
own file for organization.
