+++ frontend/src/components/module-editor/content/modals/EditImageModal.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const EditTextModal = (): React.ReactElement => <></>;
+
+export default EditTextModal;
