+++ frontend/src/components/module-editor/content/modals/index.ts
@@ -1,0 +1,4 @@
+import EditImageModal from "./EditImageModal";
+import EditTextModal from "./EditTextModal";
+
+export { EditTextModal, EditImageModal };
