+++ frontend/src/constants/DummyData.ts
@@ -1,7 +1,8 @@
 import { v4 as uuid } from "uuid";
 
 import { CourseResponse } from "../APIClients/types/CourseClientTypes";
-import { ContentTypeEnum, LessonsType } from "../types/ModuleEditorTypes";
+import { LessonsType } from "../types/ModuleEditorTypes";
+import { ContentTypeEnum } from "../types/ContentBlockTypes";
 
 export const DEFAULT_IMAGE =
   "https://res.cloudinary.com/practicaldev/image/fetch/s--JIe3p0M4--/c_limit%2Cf_auto%2Cfl_progressive%2Cq_auto%2Cw_880/https://dev-to-uploads.s3.amazonaws.com/i/093ewdrgyf1kedlhzs51.png";
